backend_secure/requirements.txt
Flask[async]==2.0.1
flask-sqlalchemy==2.5.1
psycopg2==2.9.1
pylint==2.8.2
flask-expects-json==1.6.0
Flask-Migrate==3.0.1
Jinja2==3.0.1
Flask-WTF==0.15.1
argon2-cffi==21.1.0
WTForms[email]==2.3.3
PyJWT==2.1.0